requirements.txt
alembic==1.0.7
aniso8601==4.1.0
blinker==1.4
Click==7.0
Flask==1.0.2
Flask-Mail==0.9.1
flask-marshmallow==0.9.0
Flask-Migrate==2.3.1
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==2.18.0
marshmallow-sqlalchemy==0.15.0
psycopg2==2.7.7
PyJWT==1.7.1
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
pytz==2018.9
six==1.12.0
SQLAlchemy==1.2.17
Werkzeug==0.14.1
